+++ orix/crystal_map/phase_list.py
@@ -24,9 +24,10 @@ from pathlib import Path
 from typing import Generator
 import warnings
 
-from diffpy.structure import Structure
+from diffpy.structure import Lattice, Structure
 from diffpy.structure.parsers import p_cif
 from diffpy.structure.spacegroups import GetSpaceGroup, SpaceGroup
+from diffpy.structure.symmetryutilities import ExpandAsymmetricUnit
 import matplotlib.colors as mcolors
 import numpy as np
 
@@ -48,10 +49,13 @@ ALL_COLORS.update(mcolors.XKCD_COLORS)
 class Phase:
     """Name, symmetry, and color of a phase in a crystallographic map.
 
+    If the first parameter is a :code:`Phase` instance, a copy is made
+
     Parameters
     ----------
     name
         Phase name. Overwrites the name in the ``structure`` object.
+        If this parameter is a :code:`Phase`, a copy is made.
     space_group
         Space group describing the symmetry operations resulting from
         associating the point group with a Bravais lattice, according
@@ -95,12 +99,21 @@ class Phase:
 
     def __init__(
         self,
-        name: str | None = None,
+        name: str | "Phase" | None = None,
         space_group: int | SpaceGroup | None = None,
         point_group: int | str | Symmetry | None = None,
         structure: Structure | None = None,
         color: str | None = None,
     ) -> None:
+        if isinstance(name, Phase):
+            return Phase.__init__(
+                self,
+                name.name,
+                name.space_group,
+                name.point_group,
+                name.structure.copy(),
+                name.color,
+            )
         self.structure = structure if structure is not None else Structure()
         if name is not None:
             self.name = name
@@ -134,9 +147,9 @@ class Phase:
             new_matrix = _new_structure_matrix_from_alignment(old_matrix, x="a", z="c*")
             value = copy.deepcopy(value)
             # Ensure atom positions are expressed in the new basis
-            old_xyz_cartn = value.xyz_cartn
-            value.lattice.setLatBase(new_matrix)
-            value.xyz_cartn = old_xyz_cartn
+            value.placeInLattice(Lattice(base=new_matrix))
+            # Store old lattice for expand_asymmetric_unit
+            self._diffpy_lattice = old_matrix
             if value.title == "" and hasattr(self, "_structure"):
                 value.title = self.name
             self._structure = value
@@ -352,17 +365,70 @@ class Phase:
         """
         return copy.deepcopy(self)
 
+    def expand_asymmetric_unit(self) -> Phase:
+        """Return new instance with all symmetrically equivalent atoms.
+
+        Examples
+        --------
+        >>> atoms = [Atom("Si", xyz=(0, 0, 1))]
+        >>> lattice = Lattice(4.04, 4.04, 4.04, 90, 90, 90)
+        >>> structure = Structure(atoms = atoms,lattice=lattice)
+        >>> phase = Phase(structure=structure, space_group=227)
+        >>> phase.structure
+        [Si   0.000000 0.000000 1.000000 1.0000]
+        >>> expanded = phase.expand_asymmetric_unit()
+        >>> expanded.structure
+        [Si   0.000000 0.000000 0.000000 1.0000,
+         Si   0.000000 0.500000 0.500000 1.0000,
+         Si   0.500000 0.500000 0.000000 1.0000,
+         Si   0.500000 0.000000 0.500000 1.0000,
+         Si   0.750000 0.250000 0.750000 1.0000,
+         Si   0.250000 0.250000 0.250000 1.0000,
+         Si   0.250000 0.750000 0.750000 1.0000,
+         Si   0.750000 0.750000 0.250000 1.0000]
+        """
+        if self.space_group is None:
+            raise ValueError("Space group must be set")
+
+        # Ensure atom positions are expressed in diffpy's convention
+        diffpy_structure = self.structure.copy()
+        diffpy_structure.placeInLattice(Lattice(base=self._diffpy_lattice))
+        xyz = diffpy_structure.xyz
+        diffpy_structure.clear()
+
+        eau = ExpandAsymmetricUnit(self.space_group, xyz)
+        for atom, new_positions in zip(self.structure, eau.expandedpos):
+            for pos in new_positions:
+                new_atom = copy.deepcopy(atom)
+                new_atom.xyz = pos
+                # Only add new atom if not already present
+                for present_atom in diffpy_structure:
+                    if present_atom.element == new_atom.element and np.allclose(
+                        present_atom.xyz, new_atom.xyz
+                    ):
+                        break
+                else:
+                    diffpy_structure.append(new_atom)
+
+        # This handles conversion back to correct alignment
+        out = Phase(self)
+        out.structure = diffpy_structure
+        return out
+
 
 class PhaseList:
     """A list of phases in a crystallographic map.
 
     Each phase in the list must have a unique phase id and name.
 
+    If the first parameter is a :code:`PhaseList`, a copy is made
+
     Parameters
     ----------
     phases
         A list or dict of phases or a single phase. The other
         arguments are ignored if this is passed.
+        If a :code:`PhaseList` is given, a copy is made.
     names
         Phase names. Overwrites the names in the ``structure`` objects.
     space_groups
@@ -436,7 +502,7 @@ class PhaseList:
 
     def __init__(
         self,
-        phases: Phase | list[Phase] | dict[int, Phase] | None = None,
+        phases: Phase | list[Phase] | dict[int, Phase] | "PhaseList" | None = None,
         names: str | list[str] | None = None,
         space_groups: int | SpaceGroup | list[int | SpaceGroup] | None = None,
         point_groups: str | int | Symmetry | list[str | int | Symmetry] | None = None,
@@ -445,6 +511,16 @@ class PhaseList:
         structures: Structure | list[Structure] | None = None,
     ) -> None:
         """Create a new phase list."""
+        if isinstance(phases, PhaseList):
+            return PhaseList.__init__(
+                self,
+                names=phases.names,
+                space_groups=phases.space_groups,
+                point_groups=phases.point_groups,
+                colors=phases.colors,
+                ids=phases.ids,
+                structures=phases.structures,
+            )
         d = {}
         if isinstance(phases, list):
             try:
@@ -485,7 +561,13 @@ class PhaseList:
             max_entries = max(
                 [
                     len(i) if i is not None else 0
-                    for i in [names, space_groups, point_groups, ids, structures]
+                    for i in [
+                        names,
+                        space_groups,
+                        point_groups,
+                        ids,
+                        structures,
+                    ]
                 ]
             )