+++ CHANGELOG.rst
@@ -11,14 +11,21 @@ Unreleased
 
 Added
 -----
+- Added the ``@pytest.mark.slow`` marker for slow tests.  Using ``pytest . --run-slow`` will
+  run all tests, including slow ones.
 - Can now convert Quaternions and Rotations to scipy.spatial.transform.Rotation 
   objects via Quaterion.to_scipy_rotation. This also works with Orientations 
   and Misorientations, albiet with a loss of symmetry information.
 - ``ignore_zero`` option for ``Vector3d.unique`` to allow all-zero elements, which were
   previously discarded. Discarding is still enabled by default.
+- Added copy constructors to ``Phase``, ``PhaseList`` and ``CrystalMap``.
+- Added ``Phase.expand_asymmetric_unit()`` to add all symmetrically equivalent atoms to the structure.
+- Added element-wise to indexing to ``Vector3d`` class and subclasses
 
 Changed
 -------
+- .ang files now allow optional rewriting of phase names based on elements and
+  point group, as opposed to automatically  overwriting names 
 - For speed and space reasons, vectors and quaternions are now randomly
   generated using a gaussian method as opposed to rejection-based sampling.
 - Due to dependency issues, matplotlib_scalebar as been moved from a
@@ -32,6 +39,7 @@ Deprecated
 
 Fixed
 -----
+- minor speedups to 'pytest'
 
 2025-01-01 - version 0.13.3
 ===========================